fixing deletion and disabling
making sure that one you disable an owner the whole domain is disabled
also making sure when you delete a domain all data is deleted too
also ensuring that deleting or disabling a bucket does the same for its
contents.

basically if you disable or delete a top level entity all of the storage
associated with it cascades the disable or delete.

We don't want it to be possible for a 3rd party to hijack a domain or bucket